# backend/openapi.yaml
openapi: 3.1.0
info:
  title: PastoTech API
  description: |
    REST API for PastoTech — GPS-based livestock management platform.

    ## Authentication
    All endpoints (except `/auth/login` and `/auth/register`) require a Bearer JWT token
    in the `Authorization` header.

    IoT devices authenticate using an `X-API-Key` header.

    ## Multi-tenancy
    Every request is scoped to the farm extracted from the JWT claims (`farm_id`).
    You cannot access data from other farms.

    ## Freemium limits
    Free plans allow up to 5 active animals. Exceeding this limit returns HTTP 402.
  version: 1.0.0
  contact:
    name: PastoTech Support
    url: https://pastotech.io
servers:
  - url: https://api.pastotech.io
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: User created, tokens returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      access_token: { type: string }
                      refresh_token: { type: string }
                      user: { type: object }
        '400': { description: Validation error }
        '409': { description: Email already in use }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Tokens returned
        '401': { description: Invalid credentials }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200': { description: New access token }
        '401': { description: Invalid or expired refresh token }
...